protect injection XSS

+++ backend/post-service/routes/post/delete.go
@@ -9,18 +9,25 @@ import (
 )
 
 func DeletePostHandler(c *fiber.Ctx) error {
-	postIDStr := c.Params("post_id")
+	postIDStr := c.Params("PostID")
 	postID, err := strconv.Atoi(postIDStr)
-	
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid post ID"})
 	}
 
+	// Pull Data
 	var post models.Post
 	if err := db.DB.First(&post, postID).Error; err != nil {
 		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{"error": "Post not found"})
 	}
 
+	// Check Auth
+	userID := c.Locals("UserID").(uint);
+	if post.UserID != userID {
+		return c.Status(fiber.StatusForbidden).JSON(fiber.Map{"error": "You are not allowed to edit this post"})
+	}
+
+	// Delete Data
 	if err := db.DB.Delete(&post).Error; err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Failed to delete post"})
 	}
